requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
pymongo==4.6.1
dnspython==2.4.2
streamlit==1.29.0
plotly==5.18.0
shap==0.44.0
requests==2.31.0
python-dotenv==1.0.0
seaborn==0.13.0
matplotlib==3.8.2
joblib==1.3.2